variable type set to double

+++ src/AverageUtils.java
@@ -1,10 +1,10 @@
 public class AverageUtils {
 
-    public static int findAverageOfTwoNumbers(int firstValue, int secondValue) {
+    public static double findAverageOfTwoNumbers(double firstValue, double secondValue) {
         return (firstValue + secondValue) / 2;
     }
 
-    public static int findAverageOfThreeNumbers(int firstValue, int secondValue, int thirdValue) {
+    public static double findAverageOfThreeNumbers(double firstValue, double secondValue, double thirdValue) {
         return (firstValue + secondValue + thirdValue) / 3;
     }
